pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pqccn-strongswan"
version = "0.1.0"
description = "Experiment pipeline for PQ/classic strongSwan benchmarking on constrained networks"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "IPython>=8.0",
    "matplotlib>=3.8",
    "numpy>=1.24",
    "pandas>=2.0",
    "python-on-whales>=0.71",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.scripts]
pqccn-strongswan = "pqccn_strongswan.cli.orchestrator:main"
pqccn-matrix = "pqccn_strongswan.cli.matrix:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
